+++ src/main/java/com/common/repository/QuestionRepository.java
@@ -2,15 +2,13 @@ package com.common.repository;
 
 import com.common.model.Question;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 @Repository
-public interface QuestionRepository extends JpaRepository<Question, Integer> {
+public interface QuestionRepository extends JpaRepository<Question, Long> {
 
-    List<Question> findByInterview_IdInterview(int idInterview);
+    List<Question> findByInterview_IdInterview(Long idInterview);
 
 }
