+++ src/main/java/com/common/repository/TypeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TypeRepository extends JpaRepository<TypeOfQuestion, Integer> {
+public interface TypeRepository extends JpaRepository<TypeOfQuestion, Long> {
 }
