+++ src/main/java/com/common/service/AnswerService.java
@@ -8,9 +8,9 @@ import java.util.List;
 public interface AnswerService {
 
     Answer addAnswer(Answer answer);
-    void delete(int id);
-    Answer getById(int id);
+    void delete(Long id);
+    Answer getById(Long id);
     Answer editAnswer(Answer answer);
     List<Answer> getAll();
-    List<Answer> findByIdQuestion(int id);
+    List<Answer> findByIdQuestion(Long id);
 }
