+++ src/main/java/com/common/service/InterviewService.java
@@ -8,9 +8,9 @@ import java.util.List;
 public interface InterviewService {
 
     Interview addInterview(Interview interview);
-    void delete(int id);
-    Interview getById(int id);
+    void delete(Long id);
+    Interview getById(Long id);
     Interview editInterview(Interview interview);
     List<Interview> getAll();
-    List<Interview> findByIdCreator(int id);
+    List<Interview> findByIdCreator(Long id);
 }
