+++ src/main/java/com/common/service/VariantService.java
@@ -8,9 +8,9 @@ import java.util.List;
 public interface VariantService {
 
     Variant addVariant(Variant variant);
-    void delete(int id);
-    Variant getById(int id);
+    void delete(Long id);
+    Variant getById(Long id);
     Variant editVariant(Variant variant);
     List<Variant> getAll();
-    List<Variant> findByIdQuestion(int id);
+    List<Variant> findByIdQuestion(Long id);
 }
